requirements.txt
pillow
tqdm
numpy
argparse